zvm support and multizone capability

+++ web/clui/routes/openshift.go
@@ -174,19 +174,6 @@ func (a *OpenshiftAdmin) Launch(ctx context.Context, id int64, hostname, ipaddr 
 		err = fmt.Errorf("Cluster has no built-in subnet")
 		return
 	}
-	inNet := &net.IPNet{
-		IP:   net.ParseIP(subnet.Network),
-		Mask: net.IPMask(net.ParseIP(subnet.Netmask).To4()),
-	}
-	var primaryIP net.IP
-	if ipaddr != "" {
-		primaryIP = net.ParseIP(ipaddr)
-		if !inNet.Contains(primaryIP) {
-			log.Println("Invalid IP address or not belonging to subnet")
-			err = fmt.Errorf("Invalid IP address or not belonging to subnet")
-			return
-		}
-	}
 	secgroup := &model.SecurityGroup{Model: model.Model{Owner: memberShip.OrgID}, Name: "openshift"}
 	err = db.Where(secgroup).Take(secgroup).Error
 	if err != nil {
@@ -194,7 +181,7 @@ func (a *OpenshiftAdmin) Launch(ctx context.Context, id int64, hostname, ipaddr 
 		return
 	}
 	secGroups := []*model.SecurityGroup{secgroup}
-	instance = &model.Instance{Model: model.Model{Creater: memberShip.UserID, Owner: memberShip.OrgID}, Hostname: hostname, FlavorID: flavorID, Status: "pending", ClusterID: id}
+	instance = &model.Instance{Model: model.Model{Creater: memberShip.UserID, Owner: memberShip.OrgID}, Hostname: hostname, FlavorID: flavorID, Status: "pending", ZoneID: openshift.ZoneID, ClusterID: id}
 	err = db.Create(instance).Error
 	if err != nil {
 		log.Println("DB create instance failed", err)
@@ -202,7 +189,7 @@ func (a *OpenshiftAdmin) Launch(ctx context.Context, id int64, hostname, ipaddr 
 	}
 	metadata := ""
 	var ifaces []*model.Interface
-	ifaces, metadata, err = instanceAdmin.buildMetadata(ctx, subnet, ipaddr, "", nil, nil, instance, "", secGroups)
+	ifaces, metadata, err = instanceAdmin.buildMetadata(ctx, subnet, "", "", nil, nil, instance, "", secGroups, openshift.ZoneID, openshift.ID, ipaddr)
 	if err != nil {
 		log.Println("Build instance metadata failed", err)
 		return
@@ -211,11 +198,12 @@ func (a *OpenshiftAdmin) Launch(ctx context.Context, id int64, hostname, ipaddr 
 	count := 0
 	err = db.Model(&model.Instance{}).Where("cluster_id = ? and hostname like ?", id, "%worker%").Count(&count).Error
 	if err != nil {
-		log.Println("Failed to query cluster instances", err)
 		return
 	}
 	openshift.WorkerNum = int32(count)
-	control := fmt.Sprintf("inter= cpu=%d memory=%d disk=%d network=%d", flavor.Cpu, flavor.Memory*1024, flavor.Disk*1024*1024, 0)
+	rcNeeded := fmt.Sprintf("cpu=%d memory=%d disk=%d network=%d", flavor.Cpu, flavor.Memory*1024, flavor.Disk*1024*1024, 0)
+	hyperGroup, err := instanceAdmin.getHyperGroup("", openshift.ZoneID)
+	control := "select=" + hyperGroup + " " + rcNeeded
 	command := fmt.Sprintf("/opt/cloudland/scripts/backend/oc_vm.sh '%d' '%d' '%d' '%d' '%s'<<EOF\n%s\nEOF", instance.ID, flavor.Cpu, flavor.Memory, flavor.Disk, hostname, metadata)
 	err = hyperExecute(ctx, control, command)
 	if err != nil {
@@ -322,39 +310,26 @@ func (a *OpenshiftAdmin) Update(ctx context.Context, id, flavorID int64, nworker
 	return
 }
 
-func (a *OpenshiftAdmin) Create(ctx context.Context, cluster, domain, secret, cookie, haflag, version, extIP string, nworkers int32, lflavor, mflavor, wflavor, key, subnetID int64, hostrec, bundle, registry, infrtype, sback, atbundle, icsources string) (openshift *model.Openshift, err error) {
+func (a *OpenshiftAdmin) Create(ctx context.Context, cluster, domain, secret, cookie, haflag, version, extIP string, nworkers int32, lflavor, mflavor, wflavor, key, zoneID, subnetID int64, hostrec, infrtype, sback, atbundle, icsources string) (openshift *model.Openshift, err error) {
 	memberShip := GetMemberShip(ctx)
 	db := DB()
-	openshift = &model.Openshift{
-		Model:                 model.Model{Creater: memberShip.UserID, Owner: memberShip.OrgID},
-		ClusterName:           cluster,
-		BaseDomain:            domain,
-		Status:                "creating",
-		Haflag:                haflag,
-		Version:               version,
-		Flavor:                lflavor,
-		MasterFlavor:          mflavor,
-		WorkerFlavor:          wflavor,
-		Key:                   key,
-		InfrastructureType:    infrtype,
-		StorageBackend:        sback,
-		AdditionalTrustBundle: atbundle,
-		ImageContentSources:   icsources,
-	}
-	err = db.Create(openshift).Error
-	if err != nil {
-		log.Println("DB failed to create openshift", err)
-		return
-	}
 	lbIP := ""
 	subnet := &model.Subnet{Model: model.Model{ID: subnetID}}
 	if subnetID == 0 {
 		lbIP = "192.168.91.8"
 		subnetname := openshift.ClusterName + "-sn"
 		search := cluster + "." + domain
-		subnet, err = subnetAdmin.Create(ctx, subnetname, "", "192.168.91.0", "255.255.255.0", "", "", "", "", lbIP, search, "yes", "", "", openshift.ID, memberShip.OrgID)
+		zone := fmt.Sprintf("%d", zoneID)
+		subnet, err = subnetAdmin.Create(ctx, subnetname, "", "192.168.91.0", "255.255.255.0", zone, "", "", "", "", lbIP, search, "yes", "", "", memberShip.OrgID)
 		if err != nil {
 			log.Println("Failed to create openshift subnet", err)
+			return
+		}
+		subnetIDs := []int64{subnet.ID}
+		gatewayname := openshift.ClusterName + "-gw"
+		_, err = gatewayAdmin.Create(ctx, gatewayname, "", 0, 0, subnetIDs, memberShip.OrgID, zoneID)
+		if err != nil {
+			log.Println("Failed to create gateway", err)
 			return
 		}
 	} else {
@@ -363,14 +338,30 @@ func (a *OpenshiftAdmin) Create(ctx context.Context, cluster, domain, secret, co
 			return
 		}
 	}
-	subnetIDs := []int64{subnet.ID}
-	gatewayname := openshift.ClusterName + "-gw"
-	_, err = gatewayAdmin.Create(ctx, gatewayname, "", 0, 0, subnetIDs, memberShip.OrgID)
+	openshift = &model.Openshift{
+		Model:              model.Model{Creater: memberShip.UserID, Owner: memberShip.OrgID},
+		ClusterName:        cluster,
+		BaseDomain:         domain,
+		Status:             "creating",
+		Haflag:             haflag,
+		Version:            version,
+		Flavor:             lflavor,
+		MasterFlavor:       mflavor,
+		WorkerFlavor:       wflavor,
+		SubnetID:           subnet.ID,
+		Key:                key,
+		ZoneID:             zoneID,
+		InfrastructureType: infrtype,
+		StorageBackend:     sback,
+	}
+	err = db.Create(openshift).Error
 	if err != nil {
-		log.Println("Failed to create gateway", err)
+		log.Println("DB failed to create openshift", err)
 		return
 	}
-	secgroup, err := a.createSecgroup(ctx, "openshift", "192.168.91.0/24", memberShip.OrgID)
+	preSize, _ := net.IPMask(net.ParseIP(subnet.Netmask).To4()).Size()
+	cidr := fmt.Sprintf("%s/%d", subnet.Network, preSize)
+	secgroup, err := a.createSecgroup(ctx, "openshift", cidr, memberShip.OrgID)
 	lbname := "lb"
 	keyIDs := []int64{key}
 	sgIDs := []int64{secgroup.ID}
@@ -378,15 +369,23 @@ func (a *OpenshiftAdmin) Create(ctx context.Context, cluster, domain, secret, co
 	userdata := getUserdata("ocd")
 	userdata = fmt.Sprintf("%s\ncurl -k -O '%s/misc/openshift/ocd.sh'\nchmod +x ocd.sh", userdata, endpoint)
 	parts := fmt.Sprintf("pullSecret: '%s'\n", secret)
-	if bundle != "" {
-		parts = fmt.Sprintf("%s\n%s\n", parts, bundle)
+	if atbundle != "" {
+		parts = fmt.Sprintf("%s\n%s\n", parts, atbundle)
 	}
-	if registry != "" {
-		parts = fmt.Sprintf("%s\n%s\n", parts, registry)
+	if icsources != "" {
+		parts = fmt.Sprintf("%s\n%s\n", parts, icsources)
 	}
 	encParts := base64.StdEncoding.EncodeToString([]byte(parts))
-	userdata = fmt.Sprintf("%s\n./ocd.sh '%d' '%s' '%s' '%s' '%s' '%s' '%d' '%s' '%s' '%s'<<EOF\n%s\nEOF", userdata, openshift.ID, cluster, domain, endpoint, cookie, haflag, nworkers, version, extIP, hostrec, encParts)
-	_, err = instanceAdmin.Create(ctx, 1, lbname, userdata, 1, lflavor, subnet.ID, openshift.ID, lbIP, "", nil, keyIDs, sgIDs, -1)
+	infraType := openshift.InfrastructureType
+	userdata = fmt.Sprintf("%s\n./ocd.sh '%d' '%s' '%s' '%s' '%s' '%s' '%d' '%s' '%s' '%s' '%s'<<EOF\n%s\nEOF", userdata, openshift.ID, cluster, domain, endpoint, cookie, haflag, nworkers, version, infraType, extIP, hostrec, encParts)
+	image := &model.Image{}
+	err = db.Where("hypervisor_type = ?", infraType).Take(image).Error
+	if err != nil {
+		log.Println("No valid LB image exists", err)
+		return
+	}
+	lbImg := image.ID
+	_, err = instanceAdmin.Create(ctx, 1, lbname, userdata, int64(lbImg), lflavor, subnet.ID, openshift.ID, zoneID, lbIP, "", nil, keyIDs, sgIDs, -1)
 	if err != nil {
 		log.Println("Failed to create oc first instance", err)
 		return
@@ -416,7 +415,7 @@ func (a *OpenshiftAdmin) Delete(ctx context.Context, id int64) (err error) {
 		return
 	}
 	subnet := openshift.Subnet
-	if subnet != nil {
+	if subnet != nil && subnet.Type == "internal" && subnet.Name == openshift.ClusterName+"-sn" && subnet.Network == "192.168.91.0" {
 		if subnet.Router != 0 {
 			err = gatewayAdmin.Delete(ctx, subnet.Router)
 			if err != nil {
@@ -628,9 +627,17 @@ func (v *OpenshiftView) New(c *macaron.Context, store session.Store) {
 		c.HTML(500, "500")
 		return
 	}
+	zones := []*model.Zone{}
+	err = db.Find(&zones).Error
+	if err != nil {
+		c.Data["ErrorMsg"] = err.Error()
+		c.HTML(500, "500")
+		return
+	}
 	c.Data["Flavors"] = flavors
 	c.Data["Keys"] = keys
 	c.Data["Subnets"] = subnets
+	c.Data["Zones"] = zones
 	c.HTML(200, "openshifts_new")
 }
 
@@ -666,6 +673,8 @@ func (v *OpenshiftView) Launch(c *macaron.Context, store session.Store) {
 	}
 	hostname := c.QueryTrim("hostname")
 	ipaddr := c.QueryTrim("ipaddr")
+	log.Println("------------------------------------------------------")
+	log.Println("ipaddr===" + ipaddr)
 	instance, err := openshiftAdmin.Launch(c.Req.Context(), id, hostname, ipaddr)
 	if err != nil {
 		c.JSON(500, map[string]interface{}{
@@ -693,8 +702,6 @@ func (v *OpenshiftView) Create(c *macaron.Context, store session.Store) {
 	}
 	secret := c.QueryTrim("secret")
 	hostrec := c.QueryTrim("hostrec")
-	bundle := c.QueryTrim("bundle")
-	registry := c.QueryTrim("registry")
 	nworkers := c.QueryInt("nworkers")
 	if nworkers < 2 {
 		code := http.StatusBadRequest
@@ -709,6 +716,7 @@ func (v *OpenshiftView) Create(c *macaron.Context, store session.Store) {
 	wflavor := c.QueryInt64("wflavor")
 	subnet := c.QueryInt64("subnet")
 	key := c.QueryInt64("key")
+	zone := c.QueryInt64("zone")
 	infrtype := c.QueryTrim("infrtype")
 	sback := c.QueryTrim("sback")
 	atbundle := c.QueryTrim("atbundle")
@@ -716,7 +724,7 @@ func (v *OpenshiftView) Create(c *macaron.Context, store session.Store) {
 
 	cookie := "MacaronSession=" + c.GetCookie("MacaronSession")
 	permit, err := memberShip.CheckOwner(model.Writer, "subnets", int64(subnet))
-	openshift, err := openshiftAdmin.Create(c.Req.Context(), name, domain, secret, cookie, haflag, version, extIP, int32(nworkers), lflavor, mflavor, wflavor, key, subnet, hostrec, bundle, registry, infrtype, sback, atbundle, icsources)
+	openshift, err := openshiftAdmin.Create(c.Req.Context(), name, domain, secret, cookie, haflag, version, extIP, int32(nworkers), lflavor, mflavor, wflavor, key, zone, subnet, hostrec, infrtype, sback, atbundle, icsources)
 	if err != nil {
 		if c.Req.Header.Get("X-Json-Format") == "yes" {
 			c.JSON(500, map[string]interface{}{
